Imported JSONObject .jar and created build scripts
I made a new folder called lib to hold java-json.jar which contains a library for processing JSON.
I implemented a simple proof of concept of the usage of JSONObject in CommandServer.java.
I created build scripts that include the class paths for the JSON library to compile and run CommandServer.java and TestCommandServer.java called CommandServerBuild.bash and TestCommandServerBuild.bash respectively.
Note that moving the build scripts or the lib folder with respect to each other will change the class path and the script will fail.